Make button launch the camera
