+++ OutfitGenerator/Controllers/ViewController.swift
@@ -1,0 +1,51 @@
+//
+//  ViewController.swift
+//  OutfitGenerator
+//
+//  Created by Jason on 10/14/22.
+//
+
+import UIKit
+
+class ViewController: UIViewController {
+    
+    var stylesView = Styles()
+    
+    override func loadView() {
+        stylesView.button.addTarget(self, action: #selector(launchCamera), for: .touchUpInside)
+        
+        view = stylesView
+    }
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        // Do any additional setup after loading the view.
+    }
+
+    @objc func launchCamera(sender: UIButton!) {
+        let imagePicker = UIImagePickerController()
+        imagePicker.sourceType = .camera
+        imagePicker.delegate = self
+        present(imagePicker, animated: true)
+    }
+}
+
+// MARK: - UIImagePickerControllerDelegate and UINavigationControllerDelegate methods
+
+extension ViewController: UIImagePickerControllerDelegate, UINavigationControllerDelegate {
+
+    func imagePickerControllerDidCancel(_ picker: UIImagePickerController) {
+        picker.dismiss(animated: true, completion: nil)
+    }
+
+    func imagePickerController(
+        _ picker: UIImagePickerController,
+        didFinishPickingMediaWithInfo info: [UIImagePickerController.InfoKey : Any])
+    {
+        picker.dismiss(animated: true, completion: nil)
+
+        guard let image = info[UIImagePickerController.InfoKey.originalImage] as? UIImage else {
+            return
+        }
+    }
+}
